requirements.txt
flask==2.2.3
requests==2.28.2
transformers==4.28.1
torch==2.0.0
python-dotenv==1.0.0
newspaper3k==0.2.8